backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
httpx>=0.27.0
Pillow>=10.3.0
anthropic>=0.28.0
openai>=1.30.0
